Refactor handlers by moving the Handlers class to __init__.py and removing the now redundant handlers.py. Update .gitignore to include .idea directory. Adjust analytics_handler.py to remove mplcyberpunk dependency and enhance plot styling for better visibility.
