+++ requirements.txt
@@ -3,4 +3,3 @@ gspread==5.8.0
 oauth2client==4.1.3
 pandas==2.2.3
 matplotlib==3.10.0
-mplcyberpunk==0.7.6
